+++ src/index.ts
@@ -2,5 +2,4 @@ export * from './rfi-beeline';
 export * from './gql/honeycomb.middleware';
 export * from './with-span.decorator';
 export * from './rumbleship-beeline';
-export * from './service.instrumentation';
 export * from './honeycomb.interfaces';