+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * Typesafe `RumbleshipBeeline` (+factory??), enabling tracking of fullly async service contexts
+  * `@AddToTrace` decorator
 ### Removed
 ### Changed
   * Updated tslint.json, tsconfig.json, plugins, etc to support nullish coalescing.
 ### Fixed
 ### Deprecated
+  * `src/rfi-beeline.ts`: `RFIBeeline` and `HoneycombBeelineFactory` 
 ### Security